requirements.txt
openai==1.59.8
pandas==2.2.3
datasets==3.2.0
python-dotenv